review: auth 관련 로직 utils로 분리
